# LabRestApi/src/main/resources/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API для Лабораторной информационной системы",
    "version": "2.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/patients/{id}": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Получить пациента по ID",
        "description": "Возвращает пациента по указанному идентификатору",
        "operationId": "getPatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Пациент не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Пациент успешно найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Patients"
        ],
        "summary": "Обновить пациента",
        "description": "Обновляет данные пациента по указанному идентификатору",
        "operationId": "updatePatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Пациент не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные пациента",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Пациент успешно обновлен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Patients"
        ],
        "summary": "Удалить пациента",
        "description": "Удаляет пациента по указанному идентификатору",
        "operationId": "deletePatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Пациент не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "204": {
            "description": "Пациент успешно удален"
          }
        }
      }
    },
...
    "/api/v1/patients": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Получить всех пациентов",
        "description": "Возвращает список всех пациентов. Реализована пагинация: по умолчанию page=0 & size=50",
        "operationId": "getAllPatients",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список пациентов успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Пациенты не найдены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Patients"
        ],
        "summary": "Создать нового пациента",
        "description": "Создает нового пациента на основе переданных данных",
        "operationId": "createPatient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Некорректные данные пациента",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "201": {
            "description": "Пациент успешно создан",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/patients/{patientId}/orders": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Получить заказы пациента",
        "description": "Возвращает список заказов для указанного пациента",
        "operationId": "getPatientOrders",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Пациент не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Список заказов успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/search": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Поиск пациентов по ФИО",
        "description": "Возвращает список пациентов, соответствующих указанным параметрам поиска",
        "operationId": "searchPatients",
        "parameters": [
          {
            "name": "lastName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "firstName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "middleName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Пациент не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Список пациентов успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/search-by-birthdate": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Поиск пациентов по дате рождения",
        "description": "Возвращает список пациентов, родившихся в указанную дату",
        "operationId": "searchPatientsByBirthDate",
        "parameters": [
          {
            "name": "birthDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Пациент не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Список пациентов успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/search-by-all": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Поиск пациентов по ФИО и/или дате рождения",
        "description": "Возвращает список пациентов по заданным параметрам поиска",
        "operationId": "searchPatients_1",
        "parameters": [
          {
            "name": "lastName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "firstName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "middleName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "birthDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Пациент не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Список пациентов успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orders/patient/{patientId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Получить заявки по ID пациента",
        "description": "Возвращает список заявок по ID пациента",
        "operationId": "getOrdersByPatientId",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Пациент не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Список заявок успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Сообщение об ошибке",
            "example": "Пациентов с id-404 не найдено"
          },
          "statusCode": {
            "type": "integer",
            "description": "Код ошибки",
            "format": "int32",
            "example": 404
          }
        }
      },
...
      "PatientRequestDTO": {
        "required": [
          "birthDate",
          "firstName",
          "gender",
          "lastName",
          "phoneNumber",
          "snils"
        ],
        "type": "object",
        "properties": {
          "lastName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Фамилия пациента",
            "example": "Иванов"
          },
          "firstName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Имя пациента",
            "example": "Иван"
          },
          "middleName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Отчество пациента",
            "example": "Иванович"
          },
          "birthDate": {
            "type": "string",
            "description": "Дата рождения пациента yyyy-mm-dd",
            "format": "date",
            "example": "1991-01-01"
          },
          "gender": {
            "type": "string",
            "description": "Пол пациента",
            "example": "MALE",
            "enum": [
              "MALE",
              "FEMALE"
            ]
          },
          "phoneNumber": {
            "type": "string",
            "description": "Номер телефона",
            "example": "+79333010203"
          },
          "snils": {
            "type": "string",
            "description": "Номер СНИЛС",
            "example": "213123"
          }
        },
        "description": "Запрос DTO для создания и обновления пациента"
      },
      "PatientResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID пациента",
            "format": "int64"
          },
          "lastName": {
            "type": "string",
            "description": "Фамилия пациента",
            "example": "Иванов"
          },
          "firstName": {
            "type": "string",
            "description": "Имя пациента",
            "example": "Иван"
          },
          "middleName": {
            "type": "string",
            "description": "Отчество пациента",
            "example": "Иванович"
          },
          "birthDate": {
            "type": "string",
            "description": "Дата рождения пациента yyyy-mm-dd",
            "format": "date",
            "example": "1991-01-01"
          },
          "gender": {
            "type": "string",
            "description": "Пол пациента",
            "example": "MALE",
            "enum": [
              "MALE",
              "FEMALE"
            ]
          },
          "phoneNumber": {
            "type": "string",
            "description": "Номер телефона",
            "example": "+79333010203"
          },
          "snils": {
            "type": "string",
            "description": "Номер СНИЛС",
            "example": "213123"
          }
        },
        "description": "Ответ DTO для просмотра пациентов"
      },
      "OrderResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID заявки",
            "format": "int64"
          },
          "patientId": {
            "type": "integer",
            "description": "ID пациента",
            "format": "int64"
          },
          "createdDate": {
            "type": "string",
            "description": "Дата создания заявки yyyy-MM-dd HH:mm:ss",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "description": "Статус заявки",
            "example": "CANCELED",
            "enum": [
              "REGISTERED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELED"
            ]
          },
          "comment": {
            "type": "string",
            "description": "Комментарий к заявке",
            "example": "комментарий"
          }
        },
        "description": "Ответ DTO для просмотра заявки"
      },
...
    }
  }
}